Added yellowbrick.cluster for determining the optimal number of clusters for K-Means.
Implemented yellowbrick.cluster library to perform the "elbow method" for determining the optimal number of clusters in K-Means clustering. The elbow method evaluates the distortion score for different numbers of clusters and identifies the point where the distortion score starts to decrease at a slower rate, indicating the optimal number of clusters.